+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - server: include action and event names in log output
 - server: log all HTTP errors in remote schema calls as `remote-schema-error` with details
 - console: add schema sharing
+- server: For BigQuery, make `global_select_limit` configuration optional with a default value of
+`1000`
 
 ## v2.0.1